Moved author name to AssetExternalProviderMetadataDto. This temporary fixes apidoc.
